Added  selected_storage_directory global property to allow the selected_storage_directory to be accessed from any location
Within build function of MainApplication create code to check the desired storage location and update the property. As such the storage location can be set either through the application or through the raw text file.

Added StorageLocation.txt to allow for the desired storage location to be saved through each load up of the application without it needing to be re-entered by the user, the desired stroage location MUST be set on line 3.

Updated StorageLocation Screen to allow the user to change the desired storage location for all outputs. A label at the top of the screen will display the current storage location. A reset button at the bottom right of the screen will reset the storage location to the current working directory of the application.

Both the execute button and the reset button call functions within SetStorageLocation.py.

Imported a number of functions within SetStorageLocation.py to provide necssary functionality.

Created two functions within SetStorageLocation.py to allow user to set their desired storage location or reset it to the current working directory of the script. They will then update SetStorageLocation.txt with the output, in order to save the preference between uses of the application

set result_of_ping property in NetTestingMenu.py to '' so as it is not necessary to start with a value.