RoutingIntMenu.kv
Renamed group InterfaceOrIP in both static and default menus to StaticInterfaceOrIp and DefaultInterfaceOrIP respectivly, as all four checkboxes were working in one group.

Changed all on_press to on_active to fit in with checkbox standard procedure.

Set the Interface egress method as default selection. Disabled and set Opacity = 0 for both IP route egress layouts so that they are hidden and unusable unless user switches to that setting.

RoutingIntMenu.py

Modified the functions for both Static and Default egress selections. They now refrence the layout and not the individual widgets increasing readability and efficeny.

Functions now also disable and set opacity = 0 for the uneeded input section. For example when IP egress is selected the Interface spinner, label and text input will not be visible or interactable to the user.

When widgets are disabled they also have text reset to '' to remove current user input.

SecuirtyConfMenu.kv

Modified layout of widgets to suit together on screen in an intutive manner using float layouts and position hints.

Tided up wording and text for widgets.

Set the majority of the layouts as hidden and disabled unless needed - In same manner as the static and default routes -

Configured the menu to react to user selection and display appropriate user entries. Checkboxes are linked up to functions within SecurityConfMenu.py.

For example when the user selects the Transport Method checkbox the relevant inputs and information will be displayed.

When user deselects an option the user inputs will be reset and the widgets hidden and disabled.

SecuirtyConfMenu.py

Created SecurityConfAuxVtyConLinesExecute function to read user input and configure network device dependent on selections user has made. For example if the user has only selected the login method and exec-timeout only they will be configured.

Configured numerous functions related to the checkboxes to dynamically enable and disable areas of the screen depending on if the user requires to interact with it.