Created NetScanWiresharkScreen within NetScanningAndTrafficMonMenu.kv to allow user to configure what settings they wish to run using TShark.
Created numerous functions within NetScanningAndTrafficMonMenu.py to dynamically modify NetScanWiresharkScreen so that only relevant widgets are displayed.

Completed 90% of the NetScanWireshark module to gather user inputs and start a network scan using TShark, currently untested.

Imported a number of modules for use within the NetScanWireshark function.

It should be noted that the NetScanWireshark function was imported in from an old script by Cameron Birnie and modified to suit the task.