Added a new float layout within the NetScanWiresharkScreen in NetScanningAndTrafficMonMenu.kv to dynamically inform the user about the traffic capture, it holds two labels. The first label will intially state that no capture has taken place. Once traffic capture has been performed it will then inform the user what interface was used, while the second label will inform the user where the output has been stored.
Added a new pre_enter statement in NetScanWiresharkScreen to execute the newly created function - NetScanWiresharkPopulateNetworkInterfaces() to dynamically fill the network interface spinner.

NetScanningAndTrafficMonMenu.py

Imported two modules to allow for the extraction of network interface names

In the NetScanWiresharkPopulateNetworkInterfaces function two if else statements were added to provide default entries to the capture_interface and total_duration variables in the event the user left them at default/blank.

Added fucntionality to modified the two informational labels during the execution of traffic caputre.

Used the values of netifaces.interfaces() to extract the GUID of the network interfaces. As the main OS for this script is currently Windows OS the get_connection_name_from_guid() function is called to extract the actual name from the registry.

WiresharkCaptureInterfaceSpinner is then populated with the network interfaces.

Implemented a new function get_connection_name_from_guid to extract the interface names from the GUID from the output of netifaces.interfaces(). This is done by examing the registry on windows machines. this function was provided by Gord Thompson.