Modified the order of SetStorageLocationChangeExecute function within SetStorageLocation.py to improve readablity.
Created DeviceInfoPollAndExtractScreen within DeviceInfoAndChangeControl.kv, It allows the user to select the information they wish to extract from the device, as well as if it should be stored locally or just displayed on screen.

DeviceInfoPollAndExtractExecute function within DeviceInfoAndChangeControl.py to gather user input and poll the network device for information requested by the user.

Imported and tailored a number of modules from 1st year of uni. That use re to search strings for expersions, this has been used to find the specific information has requested and return it as a variable. A number of additional functions were created using these imported modules as a base = SearchDomain and SearchUptime.

DeviceInfoPollAndExtractStoreLocalSelect function has also been created to dynamically display the widgets related to storing the output locally.