Added a label to the top of each screen that will be showcased in the Demo presentation. This label will have texted added to it once the execution of the task is complete to provide user feedback.
It will infom the user that the task has been completed and provide information such as the variable used or the IP address of the device.

In order to facilitate this in the SecurityConfAuxVtyConLinesExecute function two new variables were created - start_line_range and end_line_range. This has also allowed the line_range variable to be made more readable when dealing with a VTY selection