Added three additional screens within the DeviceInfoChangeControlMenu area -
DeviceInfoChangeControlSaveConf DeviceInfoChangeControlUploadConf DeviceInfoChangeControlCompareConf

Defined the screens as required for the functions.

Configured DeviceInfoChangeControlMenu to allow navigation to the three new submenus

DeviceInfoChangeControlSaveConf allows the user to save either a running or startup configuration from a device.

SaveDeviceRunConf and SaveDeviceStartupConf have been created as sepearte functions to allow for code reuse and increase modularity

DeviceInfoChangeControlUploadConf allows user to select a config file and upload it to a device.