Implemented functionality to Compare Configurations
Imported numerous modules to assist in performing functions -

ntpath
pathlib
webbrowser
difflib

The function makes use of difflib to compare two files, the output is then written to a HTML file which can be viewed by the user.

Updated the DeviceInfoChangeControlCompareConf screen KV rules to remove the textinput, instaed a button will be used to open up the comparison in HTMl. This is due to Kivy not supporting HTML rendering

Defined File1ChoosePopup and File2ChoosePopup KV rules to provide a popup which allows users to select files to compare

Removed redundant label from SetStorageLocation.kv