Implemented device credential entry functionality
This is done by having a button on the bottom left of every screen that connects to a device, this button will load a popup requesting the username and password. As such KVs have been updated to provide this new button.

device_username and device_password have been set as global properties in the main python file, this will allow them to be called anywhere in the script, these properties are emptied at the completion of functions

Created a new popup for when user does not enter a username or password

New file MiscModules.py has been crreated to store the DeviceUsernameAndPasswordPopup class. This allows it to be easily imported into any python file.